Messed around with the fitter. Also, making diamond maps.
Preliminary findings: Diamonds are exactly as expected based on theory. Features within the gap do *not* include a clear NDC.
NB: While it does peak at +- 0.30, the drop thereafter is insufferably slow. Does not explain the measurement.

Note that the simple nature of adding the interaction term, which automatically means e_l != e_r,
causes a current drop sufficient to explain the NDC.